Collapse application related methods into single module in lib
