--- ember_debug/lib/send-apps.js
@@ -1,14 +1,0 @@
-export default function sendApps(adapter, apps) {
-  const serializedApps = apps.map((app) => {
-    return {
-      applicationName: app.applicationName,
-      applicationId: app.applicationId,
-    };
-  });
-
-  adapter.sendMessage({
-    type: 'apps-loaded',
-    apps: serializedApps,
-    from: 'inspectedWindow',
-  });
-}
